+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2024-04-12
 
 - 客户端首页 | Client home page
+- 实现已发布内容的搜索功能 | Implement search feature for published content
 
 ## 2024-04-11